inline input

+++ Calc.go
@@ -4,21 +4,21 @@ import "strings"
 
 //Calc Adalah
 type Calc struct {
-	BilanganSatu int
-	BilanganDua  int
+	BilanganSatu int64
+	BilanganDua  int64
 	operasi      string
 }
 
-func (c *Calc) setSatu(angka int) {
+func (c *Calc) setSatu(angka int64) {
 	c.BilanganSatu = angka
 }
-func (c *Calc) setDua(angka int) {
+func (c *Calc) setDua(angka int64) {
 	c.BilanganDua = angka
 }
 func (c *Calc) setOps(str string) {
 	c.operasi = str
 }
-func (c Calc) hitung() (output int) {
+func (c Calc) hitung() (output int64) {
 	// var hasil int
 	if c.operasi == "+" {
 		output = c.BilanganSatu + c.BilanganDua
